Implement call_function
