+++ radix-engine/src/system/invocation/mod.rs
@@ -1,2 +1,4 @@
-pub mod native_wrapper;
-pub mod scrypto_executables;
+pub mod invoke_native;
+pub mod invoke_scrypto;
+pub mod resolve_function;
+pub mod resolve_method;
